♻️ Refactor: 바뀐 컴포넌트명 적용
